requirements.txt
numpy
pandas
tensorflow
pygad
py-tlsh
scikit-learn
